--- source/Coop/Mod/GameSync/Fief/FiefPatch.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using HarmonyLib;
-
-namespace Coop.Mod.GameSync
-{
-    //[HarmonyPatch(typeof(Activator), "CreateInstance")]
-    //class FiefPatch
-    //{
-    //    static bool Prefix()
-    //    {
-    //        return true;
-    //    }
-    //}
-}
